requirements.txt
hatch
pycapnp
hypothesis